rename: routing 변경사항 적용
