--- src/app/(my-repo)/me/detected-files/mock/detectedFiles.ts
@@ -1,256 +1,0 @@
-import { DetectedFilesProps } from "../_components/DetectedFile";
-
-export const DETECTED_FILES: DetectedFilesProps[] = [
-  {
-    id: 1,
-    name: "d7DwmiXWYSOKnmkUdhFTWun86PhhGNvWheh",
-    owner: { login: "홍길동" },
-    visibility: "private",
-    date: "2020-06-17",
-  },
-  {
-    id: 2,
-    name: "SlIE0r2HzibJL20WMR7JuYrOpkR2",
-    owner: { login: "정지훈" },
-    visibility: "private",
-    date: "2024-02-03",
-  },
-  {
-    id: 3,
-    name: "oSI7BKSJnEYWzQpapj0",
-    owner: { login: "정지훈" },
-    visibility: "public",
-    date: "2020-06-11",
-  },
-  {
-    id: 4,
-    name: "AHpbUiULtPTEReYh4R1bkFOXqNmG1mH9bfnnt",
-    owner: { login: "신동엽" },
-    visibility: "private",
-    date: "2024-06-03",
-  },
-  {
-    id: 5,
-    name: "w9nFTVlrLX",
-    owner: { login: "정지훈" },
-    visibility: "public",
-    date: "2023-12-15",
-  },
-  {
-    id: 6,
-    name: "wJGXPnGCuwNWbVFkSINzIit0lalrqReM",
-    owner: { login: "최유리" },
-    visibility: "private",
-    date: "2024-09-09",
-  },
-  {
-    id: 7,
-    name: "L9u2msTn87JMm04hG",
-    owner: { login: "강호동" },
-    visibility: "private",
-    date: "2021-10-03",
-  },
-  {
-    id: 8,
-    name: "XoQRZKeNtRiYEF4iN86iJydAZCW6yuJXo78DSObe",
-    owner: { login: "고은별" },
-    visibility: "private",
-    date: "2022-05-08",
-  },
-  {
-    id: 9,
-    name: "28AZBGh4yX2SYPPJNIA1VIblaKEWxHUDIPa",
-    owner: { login: "이민호" },
-    visibility: "public",
-    date: "2020-04-28",
-  },
-  {
-    id: 10,
-    name: "tM4aDCM2zFqpJdlb23iO2",
-    owner: { login: "최유리" },
-    visibility: "private",
-    date: "2024-04-28",
-  },
-  {
-    id: 11,
-    name: "0wDwsNMeDIO9o9SgigzXF01dF6V26YH",
-    owner: { login: "홍길동" },
-    visibility: "public",
-    date: "2020-08-13",
-  },
-  {
-    id: 12,
-    name: "xtWUKHIU3kLejRiGMC9oBzPg85em",
-    owner: { login: "정지훈" },
-    visibility: "private",
-    date: "2024-10-16",
-  },
-  {
-    id: 13,
-    name: "6U2OjXHRax1xgj2ibZf5Ywi7lM0b7",
-    owner: { login: "박영희" },
-    visibility: "public",
-    date: "2020-12-18",
-  },
-  {
-    id: 14,
-    name: "5XN1OiED4QNE5xI6VWIADTCZvd7AhW6Yqe",
-    owner: { login: "정지훈" },
-    visibility: "private",
-    date: "2022-12-07",
-  },
-  {
-    id: 15,
-    name: "smbJCgmUtiiWje19",
-    owner: { login: "강호동" },
-    visibility: "private",
-    date: "2020-01-09",
-  },
-  {
-    id: 16,
-    name: "eADdYmPgLdI6KYOkOAJEc",
-    owner: { login: "박영희" },
-    visibility: "private",
-    date: "2024-02-09",
-  },
-  {
-    id: 17,
-    name: "gCAdafpIvsjDav",
-    owner: { login: "홍길동" },
-    visibility: "public",
-    date: "2021-12-06",
-  },
-  {
-    id: 18,
-    name: "aSjajJEh6x3AC",
-    owner: { login: "강호동" },
-    visibility: "public",
-    date: "2024-10-25",
-  },
-  {
-    id: 19,
-    name: "WvCQE5EfW",
-    owner: { login: "홍길동" },
-    visibility: "public",
-    date: "2021-05-01",
-  },
-  {
-    id: 20,
-    name: "y91bWXji7o3fAc3WtB7c",
-    owner: { login: "이민호" },
-    visibility: "private",
-    date: "2023-06-27",
-  },
-  {
-    id: 21,
-    name: "w2eq3V3jDrecRJ5",
-    owner: { login: "이민호" },
-    visibility: "private",
-    date: "2021-01-22",
-  },
-  {
-    id: 22,
-    name: "8vIovO8vO4USNNiY",
-    owner: { login: "박영희" },
-    visibility: "private",
-    date: "2022-06-16",
-  },
-  {
-    id: 23,
-    name: "q1jMjhTNB2pgtHg",
-    owner: { login: "강호동" },
-    visibility: "public",
-    date: "2024-06-22",
-  },
-  {
-    id: 24,
-    name: "6jDj2twC5yD97DVFccML0ntmJpvpRzMoL93AGR",
-    owner: { login: "이민호" },
-    visibility: "public",
-    date: "2023-02-18",
-  },
-  {
-    id: 25,
-    name: "ZkNfvMhCWlQZc8jHWICJjSZ3V",
-    owner: { login: "홍길동" },
-    visibility: "public",
-    date: "2023-09-19",
-  },
-  {
-    id: 26,
-    name: "d0SBjXnGE6MBPKFQoQXOVw",
-    owner: { login: "김태리" },
-    visibility: "public",
-    date: "2021-05-20",
-  },
-  {
-    id: 27,
-    name: "fnUoPtf3D6M8MibvgSt3HJy7Z2pTi",
-    owner: { login: "최유리" },
-    visibility: "private",
-    date: "2022-07-22",
-  },
-  {
-    id: 28,
-    name: "NUTvBCxtKepbexbalXsWYKX7qC6kt0ifeZMnxP",
-    owner: { login: "홍길동" },
-    visibility: "public",
-    date: "2023-03-13",
-  },
-  {
-    id: 29,
-    name: "K6s1X1c5aWdMYs",
-    owner: { login: "박영희" },
-    visibility: "public",
-    date: "2023-11-24",
-  },
-  {
-    id: 30,
-    name: "YOiqF0waYkKO7XbvT",
-    owner: { login: "정지훈" },
-    visibility: "private",
-    date: "2023-10-27",
-  },
-  {
-    id: 31,
-    name: "taqzBCla",
-    owner: { login: "강호동" },
-    visibility: "private",
-    date: "2024-01-01",
-  },
-  {
-    id: 32,
-    name: "UZqHCBIzPDf6NcZOyxHUL",
-    owner: { login: "김철수" },
-    visibility: "private",
-    date: "2020-06-17",
-  },
-  {
-    id: 33,
-    name: "OhZlAKY8ORxkyZCfPl5k4",
-    owner: { login: "김태리" },
-    visibility: "public",
-    date: "2020-01-16",
-  },
-  {
-    id: 34,
-    name: "ezUpBSih1ia2hSU6Ta",
-    owner: { login: "박영희" },
-    visibility: "private",
-    date: "2024-08-04",
-  },
-  {
-    id: 35,
-    name: "Fo6QhkCBBUcw5eetYLmkZKhqK",
-    owner: { login: "이민호" },
-    visibility: "public",
-    date: "2023-02-24",
-  },
-  {
-    id: 36,
-    name: "QbIXxxpIjj",
-    owner: { login: "홍길동" },
-    visibility: "private",
-    date: "2022-11-05",
-  },
-];
