+++ src/app/me/scraps/_components/ClippingArticle.tsx
