+++ src/app/repos/_components/RepositoryItem.tsx
@@ -1,4 +1,4 @@
-import LibraryItem from "../../_components/LibraryItem";
+import LibraryItem from "./LibraryItem";
 import AssistChips from "@/components/common/chips/AssistChips";
 import Link from "next/link";
 import { RepositoryProps } from "@/types";
@@ -18,7 +18,7 @@ export default function RepositoryItem({
         </LibraryItem.Chip>
         <LibraryItem.TextBox>
           <LibraryItem.Title>{name}</LibraryItem.Title>
-          <LibraryItem.Desc>{description || 'No description'}</LibraryItem.Desc>
+          <LibraryItem.Desc>{description || "No description"}</LibraryItem.Desc>
         </LibraryItem.TextBox>
       </LibraryItem>
     </Link>
