+++ src/app/vuldb/_utils/chipsLogic.ts
